# backend_api/src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for a Medium clone",
    "version": "1.0.0",
    "title": "Medium Clone API"
  },
  "host": "localhost:3000/api/v1",
  "basePath": "/",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
  "paths": {
...
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "tags": ["Users"],
        "summary": "(Admin) Update user by ID",
        "security": [
          {
            "Bearer": [],
            "role": ["admin"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "Updated user details",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "User not found"
          }
        }
      },

      "delete": {
        "tags": ["Users"],
        "summary": "(Admin) Delete user by ID",
        "security": [
          {
            "Bearer": [],
            "role": ["admin"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
...
}